requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
djangorestframework==3.6.4
idna==2.6
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
